feat(build): 重构构建系统并完善功能测试标准
- 添加 .gitattributes 文件统一行尾规范，区分文本与二进制文件处理
- 更新 .gitignore 添加临时文件和备份文件过滤规则
- 新增 功能测试标准.md 详细文档，定义 KAT v2 功能测试标准
- 重构 build.sh 构建脚本，支持参数化构建和编译时采集功能
- 新增 field_dict.json 定义数据字段字典
- 新增 group_dict.sql 定义测试组与环境变量翻译表
- 扩展构建矩阵支持 18 个测试组和多种 ABI 形态
- 实现编译时采集阶段，支持真 x86 本地采集真值表
- 添加 --groups 参数支持指定分组构建，优化构建效率
- 完善工具链检查，增加 ar 工具验证
- 统一命名契约使用短横线分隔组与形态

+++ src/isb_cpuid.c
@@ -4,7 +4,7 @@
  * 冷启动首次 cpuid 必须在任何 cpuid 之前测(含翻译器/运行库冷初始化摊分);
  * 能力位判定复用 ib_cpu_ok(ib.h, GNU asm cpuid, mingw/Linux 通用)。
  */
-#include "ib.h"
+#include "ib_core.h"
 
 static const char *g_caps[] = {
     "sse", "sse2", "sse3", "ssse3", "sse4.1", "sse4.2", "popcnt", "movbe",
@@ -49,9 +49,9 @@ int main(int argc, char **argv)
     /* 冷启动首次 cpuid(进程内第一个 cpuid; 翻译器初始化摊分在内) */
     {
         unsigned a, b, c, d;
-        double t0 = ib_now();
+        uint64_t t0 = ib_now();
         ib_cpuid(0, 0, &a, &b, &c, &d);
-        snprintf(nb, sizeof nb, "%.1f", ib_now() - t0);
+        snprintf(nb, sizeof nb, "%.1f", (double)(ib_now() - t0));
         snprintf(dt, sizeof dt, "leaf0max=%u", a);
         ib_out("cpuid", "first_call", "diag", "OK", nb, "ns", "-", "-", "-", dt);
     }
